用户登录，登录成功后返回一个token，以后前端的每次请求都需要加上这个参数，否则会被要求重新登录

+++ src/main/java/edu/bupt/mapper/UserMapper.java
@@ -7,6 +7,6 @@ import org.springframework.stereotype.Repository;
  * Created by 73681 on 2018/5/30.
  */
 @Repository
-public interface UserMapper extends BaseMapper<User,Long>{
+public interface UserMapper extends BaseMapper<User,Integer>{
     User selectByUsername(String username);
 }
